FEATURE: Households -- added profile item to settings page. -- fixing button on households and adding chevron icon. -- creating household and household members service -- removing environments.ts from repo, updated the API key and added a secret key that requires to be kept a secret. IM me for the new cipher -- adding a way to specify if tag is apart of a household

+++ src/app/home/tags/tag-details/components/tag-detail.component.ts
@@ -5,8 +5,12 @@ import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { IonItem, IonButtons, IonButton, IonInput, IonList, IonSelect, IonSelectOption } from '@ionic/angular/standalone';
 import { NgxColorsModule } from 'ngx-colors';
-import { ReplaySubject, combineLatest, debounceTime, distinctUntilChanged, map, takeUntil, tap } from 'rxjs';
+import { ReplaySubject, Subject, combineLatest, debounceTime, distinctUntilChanged, map, takeUntil, tap } from 'rxjs';
+import { AuthService } from 'src/app/shared/_data-access/auth.service';
+import { HouseholdService } from 'src/app/shared/_data-access/household.service';
+import { SupabaseService } from 'src/app/shared/_data-access/supabase.service';
 import { TagsService } from 'src/app/shared/_data-access/tags.service';
+import { TodosService } from 'src/app/shared/_data-access/todos.service';
 import { TbItems, TbTags } from 'src/app/shared/types/database';
 
 @Component({
@@ -28,12 +32,24 @@ import { TbItems, TbTags } from 'src/app/shared/types/database';
                 formControlName="color"/>
             </ion-item>
             <ion-item>Created on {{selectedTag.created_at | date: 'shortDate'}}</ion-item>
+
+            @if(selectedTag.created_by === authService.currentUser()?.id) {
+                <ion-item>
+                    <ion-select label="Share with Household?" interface="popover" formControlName="household" required multiple>
+                        @for(household of householdService.households(); track $index) {
+                            <ion-select-option [value]="household.household_id">{{household.name}}</ion-select-option>
+                        }
+                    </ion-select>
+                </ion-item>
+            }
         </ion-list>
        
       
-        <ion-buttons>
-            <ion-button color="danger" (click)="deleteTodo(selectedTag)">Delete</ion-button>
-        </ion-buttons>
+        @if(selectedTag.created_by === authService.currentUser()?.id) {
+            <ion-buttons>
+                <ion-button color="danger" (click)="deleteTodo(selectedTag)">Delete</ion-button>
+            </ion-buttons>
+        }
     }
     `,
     styles: ``,
@@ -48,14 +64,20 @@ export class TagDetailComponent implements OnDestroy {
     private fb = inject(FormBuilder)
     private route = inject(ActivatedRoute);
     private router = inject(Router);
+    private supabaseService = inject(SupabaseService);
+
     tagsService = inject(TagsService);
+    todosService = inject(TodosService);
+    householdService = inject(HouseholdService);
+    authService = inject(AuthService);
     color ='#fff';
     textColor: 'black'|'white' = 'black';
-    destroy$ = new ReplaySubject<boolean>(1);
+    destroy$ = new Subject<boolean>();
 
     form = this.fb.nonNullable.group({
         name: ['', Validators.required],
-        color: ['', Validators.required]
+        color: ['', Validators.required],
+        household: this.fb.control<number[]>([])
     })
 
     selectedTag$ = combineLatest([
@@ -64,10 +86,12 @@ export class TagDetailComponent implements OnDestroy {
     ]).pipe(
         map(([params, tags]) => tags.find(tag => tag.tag_id === +params['id'])),
         tap(selectedTag => {
+            this.destroy$.next(true);
             if (selectedTag === undefined) return;
             this.form.patchValue({
                 name: selectedTag.name,
-                color: selectedTag.color
+                color: selectedTag.color,
+                household: selectedTag.households
             }, {emitEvent: false})
 
             this.form.get('name')?.valueChanges.pipe(debounceTime(1000), distinctUntilChanged(), takeUntil(this.destroy$)).subscribe(name => {
@@ -76,6 +100,14 @@ export class TagDetailComponent implements OnDestroy {
             this.form.get('color')?.valueChanges.pipe(takeUntil(this.destroy$)).subscribe(color => {
                 this.textColor = this.tagsService.hexToRgb(color);
                 this.tagsService.update$.next({id: selectedTag.tag_id, changes: {color} })
+            })
+            this.form.get('household')?.valueChanges.pipe(distinctUntilChanged(), takeUntil(this.destroy$)).subscribe(households => {
+                const householdValue = !!households && households.length !== 0 ? households : null
+                this.tagsService.update$.next({id: selectedTag.tag_id, changes: {households: householdValue} })
+                //TODO: convert this to a triggered event that happends on update to tb_tags
+                this.supabaseService.client.schema('public').from('tb_items').update({households: householdValue}).eq('tag_id', +selectedTag.tag_id).then(res => {
+                    console.log('updating all todos with this tag')
+                })
             })
         })
     )
